refactor: extract useDisabled hook to unify disabled handling
- Extract disabled logic into useDisabled hook
- Add type annotations for TypeScript inference
- Extract gap variable in useOffset to avoid duplication
